requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
qrcode[pil]==7.4.2
pillow==10.2.0
email-validator==2.1.0.post1
argon2-cffi==23.1.0


pyjwt
aiohttp
pydantic_settings
